backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mapmind-backend"
version = "0.1.0"
description = "Backend for the MapMind graph-first learning workspace"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "google-genai>=1.67.0,<2.0.0",
  "httpx>=0.28.0,<1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
